pyproject.toml
[project]
name = "flask-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "ijson (>=3.4.0,<4.0.0)",
    "flask-cors (>=6.0.0,<7.0.0)",
    "flask-apscheduler (>=1.13.1,<2.0.0)",
    "rarfile (>=4.2,<5.0)",
    "py7zr (>=0.22.0,<0.23.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "flask_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
